# src/AggStudentDiscounts.Api/openapi.yaml
openapi: 3.0.3
info:
  title: API Скидки студентов (карта заведений)
  description: |
    REST API для проекта интерактивной карты скидок для студентов.
    Поддерживает роли: Гость, Пользователь, Модератор, Администратор.
    Аутентификация через JWT (Bearer token).
  version: 1.0.0
servers:
  - url: https://api.example.com
    description: Продакшн сервер
  - url: http://localhost:5075
    description: Локальный сервер разработки
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
          description: Имя/никнейм
        university:
          type: string
          description: ВУЗ
        department:
          type: string
          description: Кафедра
        course:
          type: integer
          minimum: 1
          maximum: 6
        telegram:
          type: string
          nullable: true
        role:
          type: string
          enum: [guest, user, moderator, admin]
          default: user
        createdAt:
          type: string
          format: date-time
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - name
        - university
        - department
        - course
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        name:
          type: string
        university:
          type: string
        department:
          type: string
        course:
          type: integer
          minimum: 1
          maximum: 6
        telegram:
          type: string
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
          example: Bearer
        expiresIn:
          type: integer
        user:
          $ref: '#/components/schemas/User'

    # ---------- Заведение (опубликованное) ----------
    Place:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        address:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        category:
          type: string
          description: Категория заведения (кафе, магазин и т.д.)
        discount:
          type: string
          description: Описание скидки
        conditions:
          type: string
          description: Условия получения скидки
        validityPeriod:
          type: string
          nullable: true
          description: Период действия скидки (текст или даты)
        photos:
          type: array
          items:
            type: string
            format: uri
          description: Ссылки на фото (подтверждения)
        sourceUrl:
          type: string
          nullable: true
          description: Ссылка на источник информации
        votesYes:
          type: integer
          description: Количество голосов "Да"
        votesNo:
          type: integer
          description: Количество голосов "Нет"
        lastVoteDateYes:
          type: string
          format: date-time
          nullable: true
          description: Дата последнего голосования "Да"
        lastVoteDateNo:
          type: string
          format: date-time
          nullable: true
          description: Дата последнего голосования "Нет"
        status:
          type: string
          enum: [pending, published, rejected]
          description: Статус публикации (для модерации)
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Application:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        placeName:
          type: string
          description: Название заведения
        address:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        discount:
          type: string
        conditions:
          type: string
        validityPeriod:
          type: string
          nullable: true
        sourceUrl:
          type: string
          nullable: true
        photoUrls:
          type: array
          items:
            type: string
        enrichedData:
          type: object
          description: Автоматически обогащённые данные из геоисточников
          additionalProperties: true
        status:
          type: string
          enum: [pending, approved, rejected]
        rejectionReason:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateApplicationRequest:
      type: object
      required:
        - placeName
        - address
        - discount
        - conditions
        - photos
      properties:
        placeName:
          type: string
        address:
          type: string
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        discount:
          type: string
        conditions:
          type: string
        validityPeriod:
          type: string
        sourceUrl:
          type: string
        photos:
          type: array
          items:
            type: string
            format: binary
          description: Фото подтверждения (JPEG, HEIC, PNG, PDF)
    ApplicationListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Application'
        total:
          type: integer

    # ---------- Голосование ----------
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        details:
          type: object
security:
  - bearerAuth: []
paths:
...
  /api/auth/register:
    post:
      summary: Регистрация нового пользователя
      operationId: register
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Пользователь создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Неверные данные или email уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/login:
    post:
      summary: Вход в систему
      operationId: login
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Успешный вход
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Неверные учётные данные
  /api/auth/me:
    get:
      summary: Получить данные текущего пользователя
      operationId: getCurrentUser
      tags:
        - Authentication
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Данные пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Не авторизован

  # ----- Заведения (публичные) -----
...
  /api/applications:
    post:
      summary: Подать заявку на добавление заведения
      operationId: createApplication
      tags:
        - Applications
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CreateApplicationRequest'
      responses:
        '201':
          description: Заявка создана и отправлена на модерацию
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Application'
        '400':
          description: Ошибка валидации
        '429':
          description: Превышен лимит 4 заявки в сутки
        '401':
          description: Требуется авторизация

    get:
      summary: Получить список заявок текущего пользователя (личный кабинет)
      operationId: getUserApplications
      tags:
        - Applications
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, approved, rejected]
        - in: query
          name: page
          schema:
            type: integer
      responses:
        '200':
          description: Список заявок пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationListResponse'
  /api/applications/{id}:
    get:
      summary: Получить детали заявки
      operationId: getApplicationById
      tags:
        - Applications
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Данные заявки
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Application'
        '403':
          description: Доступ только к своим заявкам (или модератор)

    delete:
      summary: Отменить заявку (только если статус "на проверке")
      operationId: cancelApplication
      tags:
        - Applications
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Заявка отменена
        '403':
          description: Заявка принадлежит другому пользователю
        '409':
          description: Заявку нельзя отменить (уже обработана)
        '404':
          description: Заявка не найдена

  # ----- Администрирование и модерация -----
  /api/admin/applications:
    get:
      summary: Получить очередь заявок на проверке
      operationId: getModerationQueue
      tags:
        - Moderation
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            default: pending
        - in: query
          name: page
          schema:
            type: integer
      responses:
        '200':
          description: Список заявок для модерации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationListResponse'
        '403':
          description: Недостаточно прав (требуется роль модератора или админа)
  /api/admin/applications/{id}/approve:
    put:
      summary: Утвердить заявку (опубликовать заведение)
      operationId: approveApplication
      tags:
        - Moderation
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Заявка утверждена, заведение опубликовано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Place'
        '403':
          description: Недостаточно прав
  /api/admin/applications/{id}/reject:
    put:
      summary: Отклонить заявку с указанием причины
      operationId: rejectApplication
      tags:
        - Moderation
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
              required:
                - reason
      responses:
        '200':
          description: Заявка отклонена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Application'
        '403':
          description: Недостаточно прав
  /api/admin/applications/{id}:
    put:
      summary: Редактировать заявку перед утверждением
      operationId: editApplication
      tags:
        - Moderation
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                placeName:
                  type: string
                address:
                  type: string
                discount:
                  type: string
                conditions:
                  type: string
                validityPeriod:
                  type: string
                sourceUrl:
                  type: string
      responses:
        '200':
          description: Заявка обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Application'
  /api/admin/places/{id}:
    put:
      summary: Редактировать опубликованное заведение
      operationId: editPublishedPlace
      tags:
        - Moderation
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Place'
      responses:
        '200':
          description: Заведение обновлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Place'
        '403':
          description: Недостаточно прав

  # ----- Геокодинг и автоподсказки -----
...